AGENTS REVAMP 15 : Split BYO connection state into a hook for web-complexity budgets
